+++ lib/karafka/workers/base_worker.rb
@@ -2,20 +2,24 @@ module Karafka
   # Module encapsulating code related to Sidekiq workers logic
   module Workers
     # Worker wrapper for Sidekiq workers
-    class BaseWorker < ::SidekiqGlass::Worker
+    class BaseWorker
+      include Sidekiq::Worker
+      prepend WorkerGlass::Timeout
+      prepend WorkerGlass::Reentrancy
+
       attr_accessor :params, :controller_class_name
 
       # Executes the logic that lies in #perform Karafka controller method
       # @param controller_class_name [String] descendant of Karafka::BaseController
       # @param params [Hash] params hash that we use to build Karafka params object
-      def execute(controller_class_name, params)
+      def perform(controller_class_name, params)
         self.controller_class_name = controller_class_name
         self.params = params
         Karafka.monitor.notice(self.class, params: params)
         controller.perform
       end
 
-      # What action should be taken when execute method fails
+      # What action should be taken when perform method fails
       # @param controller_class_name [Class] descendant of Karafka::BaseController
       # @param params [Hash] params hash that we use to build Karafka params object
       def after_failure(controller_class_name, params)