+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Readme updates
 - Renamed cluster to actor_cluster for method names
 - #43 - wrong constant name
+- Replaced SidekiqGlass with generic WorkerGlass lib
+- #44 - Method name conflict
 
 ## 0.3.2
 - Karafka::Params::Params lazy load merge keys with string/symbol names priorities fix